use service in entity randomizer
